requirements.txt
fastapi
uvicorn
pydicom
pandas
numpy
opencv-python-headless
tensorflow-cpu==2.16.1
python-multipart
